tests: move project tests to test_projects
